# api/openapi.yaml
openapi: 3.0.3
info:
  title: AI 角色扮演互动平台 API
  version: 0.1.0
servers:
  - url: http://localhost:8080
paths:
  /api/health:
    get:
      summary: Health check
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: string
  /api/roles:
    get:
      summary: 列出可用角色
      responses:
        '200':
          description: 角色列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Role'
  /api/roles/{id}:
    get:
      summary: 获取角色详情
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 角色信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '404':
          description: 未找到
  /api/chat:
    post:
      summary: 发送文本消息并获取角色回复
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          description: 聊天回复
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
  /api/chat/stream:
    post:
      summary: SSE 流式聊天（占位）
      responses:
        '200':
          description: text/event-stream
  /api/asr:
    post:
      summary: 语音识别（占位）
      responses:
        '200':
          description: 识别文本
          content:
            application/json:
              schema:
                type: object
                properties:
                  text:
                    type: string
  /api/tts:
  /api/auth/login:
    post:
      summary: 用户登录（占位，返回示例令牌）
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
  /api/conversations:
    get:
      summary: 列出会话（占位）
      responses:
        '200':
          description: OK
  /api/messages:
    get:
      summary: 列出消息（占位）
      parameters:
        - in: query
          name: conversationId
          schema:
            type: string
      responses:
        '200':
          description: OK
components:
  schemas:
    post:
      summary: 语音合成（占位）
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  type: string
      responses:
        '200':
          description: 返回音频 URL（占位）
          content:
            application/json:
              schema:
                type: object
                properties:
                  audioUrl:
                    type: string
components:
  schemas:
    Role:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        category:
          type: string
        avatarUrl:
          type: string
        description:
          type: string
    ChatRequest:
      type: object
      required: [roleId, text]
      properties:
        roleId:
          type: string
        text:
          type: string
    ChatResponse:
      type: object
      properties:
        text:
          type: string
        audioUrl:
          type: string
          nullable: true
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string